# stock-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stock Management API",
        "description": "API de gestion de stock multi-vendor et multi-entrepôt",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/stocks": {
            "get": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Liste tous les stocks",
                "operationId": "baa79954a3882c15341f4d23df49a5f2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items par page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filtrer par entrepôt",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filtrer par produit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "low_stock",
                        "in": "query",
                        "description": "Stock bas uniquement",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des stocks récupérée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/in": {
            "post": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Entrée de stock",
                "operationId": "dd63c09f0813614d9796fc4f37d3cc23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "warehouse_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "cost_price": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "PO-001"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Réception commande fournisseur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock ajouté avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/out": {
            "post": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Sortie de stock",
                "operationId": "a2635eed304c56d5016f1c2033f4c239",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "warehouse_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "cost_price": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "SO-001"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Vente client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock retiré avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/transfer": {
            "post": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Transfert de stock entre entrepôts",
                "operationId": "6aa8a2fa4c3323335564edae06860545",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "from_warehouse_id",
                                    "to_warehouse_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "from_warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "TRF-001"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Transfert inter-entrepôts"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock transféré avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/adjust": {
            "post": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Ajustement de stock (inventaire)",
                "operationId": "878d0a4650780234015e39ad1caab42e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "warehouse_id",
                                    "new_quantity",
                                    "notes"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "new_quantity": {
                                        "type": "integer",
                                        "example": 95
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "ADJ-001"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Ajustement suite inventaire physique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock ajusté avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/movements": {
            "get": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Historique des mouvements de stock",
                "operationId": "81cf5eabc7911b2c27972a732f00840f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items par page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filtrer par produit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filtrer par entrepôt",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Date début",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Date fin",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/low-stock": {
            "get": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Produits en stock bas",
                "operationId": "672dbbeab732f8eda43c9ef419508282",
                "responses": {
                    "200": {
                        "description": "Liste des produits en stock bas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/product/{product_id}": {
            "get": {
                "tags": [
                    "Stock Management"
                ],
                "summary": "Détails du stock d'un produit",
                "operationId": "b17f811596358253b5e1f19eb12efd85",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du stock"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}